tried to generalize many different org file interactions
